Continued work on Managers
I started building the save and load methods in the respective managers
and added a count property to the Party class. It feels like I am having
to rebuild lists in functionality because I keep needing list functions
not given to the party class. Also, I will need all the fields and
properties that the Player class will use to be written before I can
make more progress.